Implemented naive Dijkstra

+++ backend/src/main/java/de/ecotram/backend/entity/network/Traversable.java
@@ -17,7 +17,7 @@ import java.util.Set;
 public abstract class Traversable extends EntityBase {
     @Getter
     @Setter
-    private float length;
+    private int length;
 
     @Getter
     @Setter
